小数点まで保存できるように変更

+++ db/migrate/20201204081152_create_body_weights.rb
@@ -1,7 +1,7 @@
 class CreateBodyWeights < ActiveRecord::Migration[6.0]
   def change
     create_table :body_weights do |t|
-      t.integer    :weight, null: false
+      t.float      :weight, null: false
       t.date       :day,    null: false
       t.references :user,   foreign_key: true
       t.timestamps
